__project4/requirements.txt
asgiref==3.8.1
certifi==2023.11.17
chardet==3.0.4
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework==3.15.2
EbookLib==0.18
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.1.5
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
lxml==5.1.0
markdown2==2.4.13
PyPDF2==3.0.1
python-docx==1.1.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
typing_extensions==4.9.0
